backend/requirements.txt
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pip==21.2.2
psycopg2-binary==2.9.2
python-dotenv==0.19.2
setuptools==58.0.4
SQLAlchemy==1.4.27
waitress==2.0.0
Werkzeug==2.0.2
wheel==0.37.0
wincertstore==0.2
